requirements.txt
langchain
langchain-openai
requests
python-dotenv
pytest